fix(controls): Fix bug in controls parsing and swap around some functions and files
